Added ability to shutdown SITs with a lot less error messages from the ServiceUnderTest and the ServiceImpersonators

+++ src/main/java/com/sixt/service/framework/registry/consul/RegistrationManager.java
@@ -100,15 +100,15 @@ public class RegistrationManager implements Runnable {
 
     @Override
     public void run() {
-        double sleepDuration = (double)1000;
+        long sleepDuration = 1000;
         while (! isRegistered.get()) {
             try {
                 attemptRegistration();
                 if (isRegistered.get()) {
                     break;
                 }
-                sleeper.sleepNoException((long)sleepDuration);
-                sleepDuration = sleepDuration * 1.5;
+                sleeper.sleepNoException(sleepDuration);
+                sleepDuration = (long) (sleepDuration * 1.5);
             } catch (Exception ex) {
                 logger.warn("Caught exception attempting service registration", ex);
             }
